separate autocomplete logic
